renamed resource package to controller
